Extract file indexer shebang detection partial
